artemis_backend/requirements.txt
flask
flask-wtf
flask-migrate
flask-sqlalchemy
flask-login
flask-cors
werkzeug
python-dotenv
PyJWT
requests